# internal/app/route.go
package app

import (
	"encoding/json"
	_ "github.com/ereminiu/filmoteka/docs"
	v1 "github.com/ereminiu/filmoteka/internal/controller"
	"github.com/ereminiu/filmoteka/internal/db"
	"github.com/sirupsen/logrus"
	httpSwagger "github.com/swaggo/http-swagger"
	"net/http"
)

func NewRouter(repos *db.Repositories, mig *db.Migrator) *http.ServeMux {
	router := http.NewServeMux()

	migrateRouter := v1.NewMigrateRouter(mig)
	authRouter := v1.NewAuthRouter(repos.Authorization)
	movieRouter := v1.NewMovieRouter(repos.Movie)
	actorRouter := v1.NewActorRouter(repos.Actor)

	// Migrations
	router.HandleFunc("POST /migrate-up", migrateRouter.MigrateUp)
	router.HandleFunc("POST /migrate-down", migrateRouter.MigrateDown)
	router.HandleFunc("POST /migrate-force", migrateRouter.MigrateForce)

	// Authorization
	router.HandleFunc("POST /sign-up", authRouter.CreateUser)
	router.HandleFunc("POST /sign-in", authRouter.GenerateToken)

	// User group
	router.HandleFunc("GET /actor-list", actorRouter.GetAllActors)
	router.HandleFunc("POST /search-movies", movieRouter.SearchMovie)
	router.HandleFunc("POST /movie-list", movieRouter.GetAllMovies)

	// Admin group
	router.HandleFunc("POST /add-actor", authRouter.UserIdentity(actorRouter.AddActor))
	router.HandleFunc("POST /add-actor-to-movie", authRouter.UserIdentity(movieRouter.AddActorToMovie))
	router.HandleFunc("PUT /change-actor-field", authRouter.UserIdentity(actorRouter.ChangeField))
	router.HandleFunc("DELETE /delete-actor", authRouter.UserIdentity(actorRouter.DeleteActor))
	router.HandleFunc("POST /add-movie", authRouter.UserIdentity(movieRouter.AddMovie))
	router.HandleFunc("DELETE /delete-movie", authRouter.UserIdentity(movieRouter.DeleteMovie))
	router.HandleFunc("DELETE /delete-movie-field", authRouter.UserIdentity(movieRouter.DeleteField))
	router.HandleFunc("PUT /change-movie-field", authRouter.UserIdentity(movieRouter.ChangeField))

	router.HandleFunc("POST /hello", getReverseName)
	router.Handle("/swagger/*", httpSwagger.Handler(httpSwagger.URL("http://localhost:3000/swagger/doc.json")))

	return router
}

type userInput struct {
	Name string `json:"name"`
	Age  int    `json:"age"`
}

// @Summary Get Reversed Name
// @Description reversing the given name
// @Accept json
// @Produce json
// @Param input body userInput true "username and age"
// @Success 200 {object} userInput
// @Failure 400 {string} string
// @Failure 500 {string} string
// @Router /hello [post]
func getReverseName(w http.ResponseWriter, r *http.Request) {
	var input userInput
	decoder := json.NewDecoder(r.Body)
	if err := decoder.Decode(&input); err != nil {
		logrus.Println(err)
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	output := userInput{Name: v1.ReverseName(input.Name), Age: input.Age + 1}
	jsonResponse, err := json.Marshal(output)
	if err != nil {
		logrus.Println(err)
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	w.Write(jsonResponse)
}

func empty(w http.ResponseWriter, r *http.Request) {
}